# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Recruitment Registry",
        "license": {
            "name": "Apache 2.0"
        }
    },
    "basePath": "/api/v1.0",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/surveys": {
            "x-swagger-router-controller": "survey.controller",
            "post": {
                "summary": "Creates a new survey",
                "operationId": "createSurvey",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "newSurvey",
                    "in": "body",
                    "description": "Definition of a new survey.  Existing questions have 'id' only and new questions have other fields.",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/newSurvey"
                    }
                }, {
                    "name": "parent",
                    "in": "query",
                    "description": "Parent id",
                    "required": false,
                    "type": "integer"
                }],
                "responses": {
                    "201": {
                        "$ref": "#/responses/idSuccess"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "get": {
                "summary": "Lists all surveys",
                "description": "Only returns id and name fields.",
                "operationId": "listSurveys",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "required": ["id", "name"],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/surveys/{id}": {
            "x-swagger-router-controller": "survey.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets the survey",
                "operationId": "getSurvey",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successul response",
                        "schema": {
                            "$ref": "#/definitions/survey"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "delete": {
                "summary": "Deletes the survey",
                "operationId": "deleteSurvey",
                "security": [{
                    "admin": []
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/surveys/text/{language}": {
            "x-swagger-router-controller": "survey.controller",
            "patch": {
                "summary": "Patches name of the survey",
                "operationId": "updateSurveyText",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "path",
                    "description": "Language code",
                    "required": true,
                    "type": "string",
                    "minLength": 2
                }, {
                    "name": "textUpdate",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/surveyTextUpdate"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successul response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/surveys/name/{name}": {
            "x-swagger-router-controller": "survey.controller",
            "get": {
                "summary": "Get named survey information",
                "operationId": "getSurveyByName",
                "parameters": [{
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successul response",
                        "schema": {
                            "$ref": "#/definitions/survey"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/surveys/answered/name/{name}": {
            "x-swagger-router-controller": "survey.controller",
            "get": {
                "summary": "Get named survey with answers",
                "operationId": "getAnsweredSurveyByName",
                "security": [{
                    "participant": []
                }],
                "parameters": [{
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successul response",
                        "schema": {
                            "$ref": "#/definitions/survey"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
        "/questions": {
            "x-swagger-router-controller": "question.controller",
            "post": {
                "summary": "Creates a new question",
                "operationId": "createQuestion",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "newQuestion",
                    "in": "body",
                    "description": "Question object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/newQuestion"
                    }
                }, {
                    "name": "parent",
                    "in": "query",
                    "description": "Parent id",
                    "required": false,
                    "type": "integer"
                }],
                "responses": {
                    "201": {
                        "$ref": "#/responses/idSuccess"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "get": {
                "summary": "Gets all questions",
                "operationId": "listQuestions",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language requested",
                    "required": false,
                    "type": "string",
                    "minLength": 1
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/questions"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/questions/{id}": {
            "x-swagger-router-controller": "question.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets the question",
                "operationId": "getQuestion",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language requested",
                    "required": false,
                    "type": "string",
                    "minLength": 1
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/question"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "delete": {
                "summary": "Deletes the question",
                "operationId": "deleteQuestion",
                "security": [{
                    "admin": []
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/questions/text/{language}": {
            "x-swagger-router-controller": "question.controller",
            "patch": {
                "summary": "Updates or translates question text fields",
                "operationId": "updateQuestionText",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "path",
                    "description": "Language code",
                    "required": true,
                    "type": "string",
                    "minLength": 2
                }, {
                    "name": "textUpdate",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/questionTextUpdate"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successul response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "admin": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        },
        "participant": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        },
        "self": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "responses": {
...
        "genericError": {
            "description": "Error response",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
...
        "idSuccess" : {
            "description": "Successful response",
            "schema": {
                "description": "Id",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "questionType": {
            "type": "string",
            "enum": ["text", "bool", "choice", "choices"]
        },
        "newSurveyQuestion": {
            "type": "object",
            "description": "Describes either a brand new question (no 'id') or an existing question ('id' only with 'required').",
            "required": ["required"],
            "properties": {
                "required": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/questionType"
                },
                "oneOfChoices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "minItems": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": ["text"],
                        "properties": {
                            "text": {
                                "type": "string",
                                "minLength": 1
                            },
                            "type": {
                                "type": "string",
                                "enum": ["text", "bool"]
                            }
                        }
                    },
                    "minItems": 1
                },
                "actions": {
                    "$ref": "#/definitions/newActions"
                }
            },
            "minProperties": 2,
            "additionalProperties": false
        },
        "newSurvey": {
            "type": "object",
            "required": [
                "name",
                "questions"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/newSurveyQuestion"
                    },
                    "minItems": 1
                }
            },
            "additionalProperties": false
        },
        "surveyQuestion": {
            "type": "object",
            "required": [
                "id",
                "text",
                "type",
                "required"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                     "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/questionType"
                },
                "required": {
                    "type": "boolean"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/choice"
                    },
                    "minItems": 1
                },
                "actions": {
                    "$ref": "#/definitions/actions"
                 },
                "answer": {
                    "$ref": "#/definitions/answer"
                }
            },
            "additionalProperties": false
        },
        "survey": {
            "type": "object",
            "required": [
                "id",
                "name",
                "questions"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                     "minLength": 1
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/surveyQuestion"
                    },
                    "minItems": 1
                }
            },
            "additionalProperties": false
        },
        "surveyTextUpdate": {
            "type": "object",
            "required": ["id", "name"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "choiceType": {
            "type": "string",
            "enum": ["bool", "text"]
        },
        "choice": {
            "type": "object",
            "required": [
                "id",
                "text"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                     "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/choiceType"
                }
            }
        },
        "choiceTextUpdate": {
            "type": "object",
            "required": [
                "id",
                "text"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                     "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "answer": {
            "type": "object",
            "properties": {
                "choice": {
                    "type": "integer",
                    "minimum": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": ["id"],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "boolValue": {
                                "type": "boolean"
                            },
                            "textValue": {
                                "type": "string"
                            }
                        },
                        "maxProperties": 2
                    },
                    "minItems": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                }
            },
            "minProperties": 1,
            "maxProperties": 1
        },
        "question": {
            "type": "object",
            "required": [
                "id",
                "text",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                     "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/questionType"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/choice"
                    }
                },
                "actions": {
                    "$ref": "#/definitions/actions"
                 },
                "answer": {
                    "$ref": "#/definitions/answer"
                }
            }
        },
        "questions": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/question"
            },
            "minLength": 1
        },
        "questionTextUpdate": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                     "minLength": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/choiceTextUpdate"
                    }
                },
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/actionTextUpdate"
                    }
                 }
            },
            "additionalProperties": false
        },
        "newChoice": {
            "type": "object",
            "required": ["text"],
            "properties": {
                "text": {
                    "type": "string",
                     "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/choiceType"
                }
            },
            "additionalProperties": false
        },
        "newQuestion": {
            "type": "object",
            "required": [
                "text",
                "type"
            ],
            "properties": {
                "text": {
                    "type": "string",
                    "minLength": 1
                },
                "type": {
                    "$ref": "#/definitions/questionType"
                },
                "oneOfChoices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/newChoice"
                    }
                },
                "actions": {
                    "$ref": "#/definitions/newActions"
                }
            },
            "additionalProperties": false
        },
...
        "newAction": {
            "type": "object",
            "required": ["text", "type"],
            "properties": {
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            },
            "additionalProperties": false
        },
        "action": {
            "type": "object",
            "required": ["id", "text", "type"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                },
                "type": {
                    "type": "string",
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "actionTextUpdate": {
            "type": "object",
            "required": ["id", "text"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "newActions": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/newAction"
            }
        },
        "actions": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/action"
            }
        },
...
    }
}